Added breaks to SteinerTree switches to emphasise that switches will not fall through.
For the modified cases, the switch statements would fall through if no return path was hit.
Although the assert is fired when no return is performed anyway, this isn't clear to the
compiler, and a fallthrough warning (thus failure due to -Werror) occurs.

Adding these breaks makes the intent clear to the compiler and avoids such issues.