First steps towards building pytket with Nix.
The concept is somewhat different to the original intent of setup.py,
simply due to keeping Nix in control of the build. Nix, rather than
setup.py, compiles binders/. It then passes that - along with python
package dependencies such as numpy etc - to the setup process. It
also passes a flag: USE_NIX=1

The setup process now detects USE_NIX and chooses a NixBuild build_ext
to put the pieces together, rather than the previous options of
CMake or Conan. In the absense of USE_NIX, the setup process is unchanged.

Whether or not "doing nothing" is the appropriate form of NixBuild,
I do not yet know. Probably not. But until we get to that stage,
we need to solve another issue: `types-pkg_resources` is a listed
dependency, and we aren't providing it. It isn't something that's in
nixpkgs, and I'm struggling to find any useful information about
it online.