Further progress on building pytket with nix.
Added nix support for types-pkg_resources via fetchPypi,
and established build of pytket. At current, tests cause
permission errors so are disabled. This must be remedied
in future, or otherwise provided in a separate derivation
such that tests can be run.

At current, no confirmation that the build is successful
and that libraries are available to pytket when running.
In particular, at the moment I'm making .a files, rather
than .so files, available for tket and libs, which may
or may not work. If it doesn't work, an alternative build
will need to be added.