Add version into setup call
The core pytket project uses setuptools_scm to extract
version information from .git. This would be awkward in
Nix as it would involve copying the .git directory into
the pytket source.

As we already have the version extracted from another
file in the project, which is used to set the version
number on the Nix side, we can instead hardcode this
version number into setup.py when copying it to the
derivation source.