Update readme with cachix usage instructions
When this PR is merged, build artifacts will be made available on
Cachix. Nix does not make this as friendly as it could be, because
accepting a cache has security implications for end users. As such,
users need to be in the 'trusted-users' list. Unfortunately, installing
a flake from a non-trusted user will still ask for permission to add the
cache, but the user's selection will be ignored by the Nix backend.

As such, instructions for allowing the tket.cachix.org cache are
provided in the README to avoid unnecessary building on end users'
machines.