Updated pytket's CMake build and nix build wrapper
Previously Nix had its own special build, but with the changes
to tket's CMake this is no longer necessary. As such, Nix now
uses the CMake build path, which has only had to change in one
manner: ./build had to be renamed to ./cmake_build, so as not
to remove python files previously placed there by setup.py.