Increasing compatibility with existing setup
Previously, the build directory being CWD/build caused problems in the
pytket setup.py process in a Nix build. Changing this to CWD/cmake_build
fixed the issue temporarily. Now I am defaulting back to CWD/build, but
allowing another directory to be provided with a BUILD_DIR environment
variable. This means the setup should be, for all intents and purposes,
identical on non-nix builds.