Updated cmake for tests
Previously tests were compiled into executables, but not registered
with ctest using catch's catch_discover_tests. I've added that in,
and added enable_testing to tket's CMakeLists.txt in the presence of
test flags.

This also allows for a simplification of the nix derivation - TKET
and its tests are now built as one project, then checked.